refactor: put colour manipulations into its own file
